Update MessageContext.cs
Bug fix for Issue #1 in which, when attempting to retrieve XElement in ToContextX12(), EdiElements.StTagHipaa is "D_ST01" but the public field as defined in Definitions/005010_834.cs is "D_ST01_TransactionSetIdentifierCode", causing an incorrect null assignment. Hard-coding the full field name fixes the problem so the incorrect name is verified to be the problem. This fix searches for any XElements in which the name simply begins with either EdiElements.StTag or EdiElements.StTagHipaa.